--- models/yesterday.js
@@ -1,15 +1,0 @@
-function getYesterdayDate() {
-  const currentDate = new Date();
-  const yesterday = new Date(currentDate);
-  yesterday.setDate(currentDate.getDate() - 1);
-
-  const year = yesterday.getFullYear();
-  const month = String(yesterday.getMonth() + 1).padStart(2, '0');
-  const day = String(yesterday.getDate()).padStart(2, '0');
-
-  return `${year}${month}${day}`;
-}
-
-module.exports = {
-  getYesterdayDate
-};
